requirements.txt
Django==2.2.7
Flask==1.1.1
gunicorn==20.0.4
gevent==1.4.0
itsdangerous==1.1.0
Jinja2==2.10.3
keras-tqdm==2.0.1
MarkupSafe==1.1.1
matplotlib==3.1.2
numpy==1.18.1
pandas==0.25.3
plotly==4.4.1
pillow==7.0.0
scikit-learn==0.22
scipy==1.4.1
sklearn==0.0
tensorflow==1.14.0
tqdm==4.41.1
Werkzeug==0.16.0

